id type change from Long to Int

+++ src/main/kotlin/com/gdsc/boolpyeon/user/domain/dto/request/UserModifyRequest.kt
@@ -2,7 +2,7 @@ package com.gdsc.boolpyeon.user.domain.dto.request
 
 // TODO( Front에서 null을 줄 수 있는가? )
 data class UserModifyRequest(
-    val id: Long,
+    val id: Int,
     val name: String?,
     val mail: String?,
     val phoneNumber: String?,
